begin

+++ SpaceAtlas.BL/Planet/PlanetService.cs
@@ -1,5 +1,6 @@
 using System.Data;
 using AutoMapper;
+using SpaceAtlas.BL.Exceptions;
 using SpaceAtlas.BL.Planet.Entities;
 using SpaceAtlas.DataAccess.Entities;
 using SpaceAtlas.DataAccess.Repository;
@@ -17,9 +18,9 @@ public class PlanetService : IPlanetService
         _planetRepository = planetRepository;
         _mapper = mapper;
     }
-    public IEnumerable<PlanetModel> GetAll(FilterPlanetModel? filter = null)
+    public async Task<IEnumerable<PlanetModel>> GetAll(FilterPlanetModel? filter = null)
     {
-        var planets = _planetRepository.GetAll(x =>
+        var planets = await _planetRepository.GetAll(x =>
             filter == null ||
             (filter.Name == null || x.Name == filter.Name) ||
             (filter.UserId == null || x.UserId == filter.UserId) ||
@@ -28,30 +29,34 @@ public class PlanetService : IPlanetService
         return _mapper.Map<IEnumerable<PlanetModel>>(planets);
     }
 
-    public PlanetModel GetById(Guid planetId)
+    public async Task<PlanetModel> GetById(Guid planetId)
     {
-        var planet = _planetRepository.GetById(planetId);
+        var planet = await _planetRepository.GetById(planetId);
         if(planet == null)
             throw new KeyNotFoundException();
         return _mapper.Map<PlanetModel>(planet);
     }
 
-    public Guid Create(PlanetModel planet)
+    public async Task<Guid> Create(PlanetModel planet)
     {
         planet.Id = Guid.NewGuid();
         var planetEntity = _mapper.Map<PlanetEntity>(planet);
-        return _planetRepository.Save(planetEntity);
+        return await _planetRepository.Save(planetEntity);
     }
 
-    public Guid Update(PlanetModel planet)
+    public async Task<Guid> Update(PlanetModel planet)
     {
         if (planet.Id != null)
         {
-            var planetEntity = _planetRepository.GetById(planet.Id.Value);
+            var planetEntity = await _planetRepository.GetById(planet.Id.Value);
             if (planetEntity == null)
-                throw new KeyNotFoundException();
+                throw new KeyNotFoundException("Такой планеты не существует");
+            if(planet.UserId == null)
+                planet.UserId = planetEntity.UserId;
+            else if (planet.UserId != planetEntity.UserId)
+                throw new IncorrectAuthorException("У вас нет прав на редактирование данной статьи");
             planetEntity = _mapper.Map<PlanetEntity>(planet);
-            return _planetRepository.Save(planetEntity);
+            return await _planetRepository.Save(planetEntity);
         }
         else
         {
@@ -59,8 +64,20 @@ public class PlanetService : IPlanetService
         }
     }
 
-    public void Delete(Guid planetId)
+    public async Task Delete(Guid planetId)
     {
-        _planetRepository.Delete(planetId);
+        var planetEntity = await _planetRepository.GetById(planetId);
+        if (planetEntity == null)
+            throw new KeyNotFoundException("Такой звезды не существует");
+        await _planetRepository.Delete(planetId);
+    }
+    public async Task Delete(Guid planetId, Guid userId)
+    {
+        var planetEntity = await _planetRepository.GetById(planetId);
+        if (planetEntity == null)
+            throw new KeyNotFoundException("Такой звезды не существует");
+        if (planetEntity.UserId != userId)
+            throw new IncorrectAuthorException("У вас нет прав на редактирование данной статьи");
+        await _planetRepository.Delete(planetId);
     }
 }
